+++ symfony/src/Api/Console/ControllerOrg/BlogController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Api\Console\Controller;
+namespace App\Api\Console\ControllerOrg;
 
 use App\Api\Console\Authorization\OrganizationLevelEndpoint;
 use App\Api\Console\Authorization\OrganizationOptional;
